universal7880: Correct extra Samsung GPS field
int automatically extends to 8 bytes on arm64.

+++ include/hardware/gps.h
@@ -571,7 +571,7 @@ typedef struct {
     /** Azimuth of SV in degrees. */
     float   azimuth;
     /** SAMSUNG */
-    uint64_t samsung;
+    int     samsung;
 } GpsSvInfo;
 
 typedef struct {
